Bug fixes found by static analysis.
This change fixes a whole bunch of bugs and minor cleanup issues
found by static analysis when run on commit
0fa780596128ef6030a8dea1863894b596f2b088.

+++ apps/libc/dynamic/wstring.c
@@ -698,7 +698,7 @@ Return Value:
 
 {
 
-    return wcsncmp(String1, String2, -1);
+    return wcsncmp(String1, String2, (size_t)-1);
 }
 
 LIBC_API
@@ -733,7 +733,7 @@ Return Value:
 
 {
 
-    return wcsncasecmp(String1, String2, -1);
+    return wcsncasecmp(String1, String2, (size_t)-1);
 }
 
 LIBC_API
